pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctl"
version = "0.1.0"
description = "Common Transformation Library (Python) with REST API"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "CTL Contributors"}]
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "pandas>=2.2",
  "numpy>=1.26",
  "typer>=0.12",
  "requests>=2.32",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.4",
]
sql = [
  "sqlalchemy>=2.0",
  "pyodbc>=5.1; platform_system=='Windows'",
  "pymssql>=2.3; platform_system!='Windows'",
  "pymysql>=1.1",
]
databricks = [
  "databricks-sql-connector>=3.2",
]
excel = [
  "xlwings>=0.31",
]
stats = [
  "statsmodels>=0.14",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100